examples/python-sdk/chat/streamlit_simple_chat/requirements.txt
streamlit==1.24.0
python-dotenv
asyncio
judini